Did Maths homework!!:). Added Aspect logging (look further into configuration). Added print decorator.

+++ src/main/java/it/hqsolutions/lastminute/exercise/bl/bo/interfaces/Receipt.java
@@ -9,11 +9,11 @@ public interface Receipt {
 
 	public List<SalableItem> getSalableItems();
 
-	public float getSalesTaxes();
+	public double getSalesTaxes();
 
-	public float getTotal();
+	public double getTotal();
 
-	public float getTotalDistinguishedItems();
+	public double getTotalDistinguishedItems();
 
 	public int addSalableItem(SalableItem salableItem);
